int[] to mimic unsigned byte[], need to check with real scope show scope name

+++ source/android/AndroidTouchScope/app/src/main/java/de/uni_weimar/mheinz/androidtouchscope/UsbController.java
@@ -103,7 +103,7 @@ public class UsbController
         return mTmcSocket.write(command);
     }
 
-    public byte[] read(int length)
+    public int[] read(int length)
     {
         if(mTmcSocket == null)
             return null;
